refactor: eventHandler 관심사 분리
